// interfaces/viewer/style.ts
export interface ViewerHeaderStyleProps {
  height: number;
}

export interface ViewerSettingColorItemStyleProps {
  isSelected: boolean;
  color: string;
}

export interface ViewerPageControllerStyleProps {
  backgroundColor?: string;
  width?: number;
  height?: number;
  menuHeight?: number;
}

export interface ViewerArticleStyleProps {
  fontSize: number;
  lineHeight: number;
}

export interface ViewerSectionStyleProps {
  width: number;
}

export interface ViewerBottomStyleProps {
  height: number;
}